lab2 separate functions to more files
