+++ lab2/get_information.h
@@ -1,0 +1,8 @@
+#ifndef GET_INFORMATION_H
+#define GET_INFORMATION_H
+
+void get_serial_number();
+void get_board_revision();
+void get_VC_memory();
+
+#endif
