+++ lab2/load_kernel.c
@@ -1,0 +1,12 @@
+#include "uart.h"
+#include "gpio.h"
+#include "mbox.h"
+#include "tools.h"
+#include "get_information.h"
+
+void load_kernel_at_target_address(const char * target_address){
+
+
+
+} 
+
