+++ lab2/load_kernel.h
@@ -1,0 +1,7 @@
+#ifndef LOAD_KERNEL_H
+#define LOAD_KERNEL_H
+
+void load_kernel_at_target_address(const char * target_address);
+
+
+#endif
