+++ lab2/main.c
@@ -23,163 +23,66 @@
  *
  */
 
+#include "tools.h"
 #include "uart.h"
 #include "mbox.h"
+#include "get_information.h"
+#include "gpio.h"
+#include "load_kernel.h"
+
 #define PM_PASSWORD (0x5a000000)
 #define PM_RSTC ((int*)0x3F10001c)
 #define PM_WDOG ((int*)0x3F100024)
 
-
-void reverse(char str[], int length) { 
-    int start = 0; 
-    int end = length -1; 
-    while (start < end) { 
-		char tmp  = *(str+start);
-		*(str+start) = *(str+end);
-		*(str+end) = tmp;
-        //swap(*(str+start), *(str+end)); 
-        start++; 
-        end--; 
-    } 
-}
-char* itoa(unsigned long long int num, char* str, unsigned long long int base) { 
-    unsigned long long int i = 0; 
-    _Bool isNegative = 0; 
-
-    if (num == 0) { 
-        str[i] = '0';
-		i = i + 1;
-        str[i] = '\0'; 
-        return str; 
-    } 
-    if (num < 0 && base == 10) { 
-        isNegative = 1; 
-        num = -num; 
-    } 
-    while (num != 0) { 
-        unsigned long long int rem = num % base; 
-        str[i++] = (rem > 9)? (rem-10) + 'a' : rem + '0'; 
-        num = num/base; 
-    }  
-    if (isNegative) 
-        str[i++] = '-'; 
-  
-    str[i] = '\0';
-    reverse(str, i); 
-    return str; 
-} 
 void main()
 {
+	// Initialization
     char buffer[100];
 	char current_input;
-    int i1, i2, i3;
-    float f1, f2, f3;
+    int i1, i2;
+    float f1, f2;
     int sizeof_current_line = 0;
     _Bool command_not_found = 1;
 	unsigned long long int tmp_CNTFRQ_EL0;
 	unsigned long long int tmp_CNTPCT_EL0;
-	unsigned long long int output_time;
 	
+	// UART0 initialization
     uart_init();
-	
-	uart_puts("\n");	
-	// get the board's unique serial number with a mailbox call
-    mbox[0] = 8*4;                  // length of the message
-    mbox[1] = MBOX_REQUEST;         // this is a request message
-    
-    mbox[2] = MBOX_TAG_GETSERIAL;   // get serial number command
-    mbox[3] = 8;                    // buffer size
-    mbox[4] = 8;
-    mbox[5] = 0;                    // clear output buffer
-    mbox[6] = 0;
+	uart_puts("\n");
+	// Print the peripheral information
+	get_serial_number();
+	get_board_revision();
+	get_VC_memory();
 
-    mbox[7] = MBOX_TAG_LAST;
-
-    // send the message to the GPU and receive answer
-
-    // get serial number
-    if (mbox_call(MBOX_CH_PROP)) {
-        uart_puts("My serial number is: ");
-        uart_hex(mbox[5]);
-        uart_hex(mbox[6]);
-		uart_puts("\n");
-    } else {
-        uart_puts("Unable to query serial!\n");
-    }
-
-	// get board revision
-	mbox[0] = 7*4;                  // length of the message
-    mbox[1] = MBOX_REQUEST;         // this is a request message 
-    mbox[2] = MBOX_TAG_GETBOARDREVISION;   // get board revision number command
-    mbox[3] = 4;                    // buffer size
-    mbox[4] = 8;
-    mbox[5] = 0;                    // clear output buffer
-    mbox[6] = MBOX_TAG_LAST;
-
-    if (mbox_call(MBOX_CH_PROP)) {
-        uart_puts("My board revision is: ");
-        uart_hex(mbox[5]);
-        uart_puts("\n");
-    } else {
-        uart_puts("Unable to query serial!\n");
-    }
-
-	// get VC Memory
-	mbox[0] = 8*4;                  // length of the message
-    mbox[1] = MBOX_REQUEST;         // this is a request message 
-    mbox[2] = MBOX_TAG_GETVCMEMORYADDR;   // get VC memory base address command
-    mbox[3] = 8;                    // buffer size
-    mbox[4] = 8;
-    mbox[5] = 0;                    // clear output buffer
-    mbox[6] = 0;
-    mbox[7] = MBOX_TAG_LAST;
-
-	if (mbox_call(MBOX_CH_PROP)) {
-        uart_puts("My VC memory base address is: ");
-        uart_hex(mbox[5]);
-		uart_puts("\n");
-		uart_puts("My VC memory size is: ");
-        uart_hex(mbox[6]);
-        uart_puts("\n");
-    } else {
-        uart_puts("Unable to query serial!\n");
-    }
-
-    // set up serial console
+    // Set up serial console
     uart_getc();
-    // say hello
-    //uart_puts("Hello World!\n");
-    uart_puts("\r\nWelcome to raspberry pi 3!\n#"); 
+    // Welcome message
+	uart_puts("\r\nWelcome to raspberry pi 3!\n# "); 
 
-    // echo everything back
-    //uart_puts("\r#");
     while(1) {
-		current_input = uart_getc();
+		current_input = uart_getc();	// read every character on keylogging
         buffer[sizeof_current_line] = current_input;
         uart_send(current_input);
 		sizeof_current_line = sizeof_current_line + 1;
         if(current_input == '\n') {
-			//uart_puts("YOU PRESS ENTER\n");
 			command_not_found = 1;
-			if(sizeof_current_line==6 && buffer[0]=='h' && buffer[1]=='e' && buffer[2]=='l' && buffer[3]=='l' && buffer[4]=='o') {
-				// print Hello World!
+			if(sizeof_current_line==6 && strcmp(buffer, "hello")) {
+				// Print Hello World!
 				uart_puts("\rHello World!\n");
 				command_not_found = 0;
 			}
-			else if(sizeof_current_line==5 && buffer[0]=='h' && buffer[1]=='e' && buffer[2]=='l' && buffer[3]=='p') {
-				// print all commands
+			else if(sizeof_current_line==5 && strcmp(buffer, "hello")) {
+				// Print all commands
 				uart_puts("\rhello: print Hello World!\n");
 				uart_puts("help: help\n");
 				uart_puts("reboot: reboot rpi3\n");
 				uart_puts("timestamp: get current timestamp\n");
 				command_not_found = 0;
 			}
-			else if(sizeof_current_line==10 && buffer[0]=='t' && buffer[1]=='i' && buffer[2]=='m' && buffer[3]=='e' && buffer[4]=='s' && buffer[5]=='t' && buffer[6]=='a' && buffer[7]=='m' && buffer[8]=='p' ) {
-				// print timestamp
-				//uart_puts("timestamp TO BE CONTINUE...\n");
+			else if(sizeof_current_line==10 && strcmp(buffer, "timestamp")) {
+				// Print timestamp
 				asm volatile("mrs %0, CNTPCT_EL0" : "=r" (tmp_CNTPCT_EL0));
 				asm volatile("mrs %0, CNTFRQ_EL0" : "=r" (tmp_CNTFRQ_EL0));
-				//output_time = tmp_CNTPCT_EL0 / tmp_CNTFRQ_EL0;
 				f1 = (float) ( (float)tmp_CNTPCT_EL0 / (float)tmp_CNTFRQ_EL0 );
 				i1 = f1;
 				f2 = f1 - i1;
@@ -192,33 +95,16 @@ void main()
 				    f2 = f2 * 10;
 				}
 
-				char buf[100];
-				char buf1[100];
-				char buf2[100];
-				char buf3[100];
-				itoa(i1, buf, 10);
-				itoa(i2, buf1, 10);
+				char buf[100], buf1[100], buf2[100], buf3[100];
+				itoa(i1, buf, 10); itoa(i2, buf1, 10);
 				itoa(tmp_CNTPCT_EL0, buf2, 10);
 				itoa(tmp_CNTFRQ_EL0, buf3, 10);
-				uart_puts("\r[");
-				uart_puts(buf);
-				uart_puts(".");
-				uart_puts(buf1);
-				uart_puts("]");
-				uart_puts("\n");
-				//uart_puts("CNTPCT_EL0: ");
-				//uart_puts(buf2);
-				//uart_puts("\n");
-				//uart_puts("CNTFRQ_EL0: ");
-				//uart_puts(buf3);
-				//uart_puts("\n");
-
+				uart_puts("\r["); uart_puts(buf); uart_puts(".");
+				uart_puts(buf1); uart_puts("]"); uart_puts("\n");
 				command_not_found = 0;
 			}
-			else if(sizeof_current_line==7 && buffer[0]=='r' && buffer[1]=='e' && buffer[2]=='b' && buffer[3]=='o' && buffer[4]=='o' && buffer[5]=='t' ) {
-				// reboot rpi3
-				//uart_puts("reboot TO BE CONTINUE...\n");
-				//uart_puts("\n\r");	
+			else if(sizeof_current_line==7 && strcmp(buffer, "reboot")) {
+				// Reboot rpi3
 
 				*PM_RSTC = ((PM_PASSWORD) | 0x20); // full reset
   				*PM_WDOG = ((PM_PASSWORD) | 10); // number of watchdog tick
@@ -227,12 +113,14 @@ void main()
 				command_not_found = 0;
 				continue;
 			}
-			else if(sizeof_current_line==12 && buffer[0]=='f' && buffer[1]=='r' && buffer[2]=='a' && buffer[3]=='m' && buffer[4]=='e' && buffer[5]=='b' && buffer[6]=='u' && buffer[7]=='f' && buffer[8]=='f' && buffer[9]=='e' && buffer[10]=='r' ) {
+			else if(sizeof_current_line==12 && strcmp(buffer, "framebuffer")) {
 				uart_puts("Show framebuffer\n");
 				command_not_found = 0;
 			}
-			else if(sizeof_current_line==8 && buffer[0]=='l' && buffer[1]=='o' && buffer[2]=='a' && buffer[3]=='d' && buffer[4]=='i' && buffer[5]=='m' && buffer[6]=='g' ) {
+			else if(sizeof_current_line==8 && strcmp(buffer, "loadimg")) {
 				uart_puts("Load kernel image\n");
+				char kernel_address_buffer[1000];
+				load_kernel_at_target_address(kernel_address_buffer);
 				command_not_found = 0;
 			}
 			if(command_not_found == 1) {
@@ -243,7 +131,7 @@ void main()
 					uart_puts("\" not found. Try <help>\n");
 				}
 			}
-			uart_puts("\r#");
+			uart_puts("\r# ");
 			sizeof_current_line = 0;
         }
