+++ lab2/mbox.h
@@ -23,6 +23,9 @@
  *
  */
 
+//#ifndef MBOX_H
+//#define MBOX_H
+
 /* a properly aligned buffer */
 extern volatile unsigned int mbox[36];
 
@@ -47,3 +50,5 @@ extern volatile unsigned int mbox[36];
 #define MBOX_TAG_LAST           		0
 
 int mbox_call(unsigned char ch);
+
+//#endif
