+++ lab2/tools.h
@@ -1,0 +1,8 @@
+#ifndef TOOLS_H
+#define TOOLS_H
+
+void reverse(char str[], int length);
+char* itoa(unsigned long long int num, char* str, unsigned long long int base);
+int strcmp(const char *X, const char *Y);
+
+#endif
